+++ CHANGES.md
@@ -1,6 +1,10 @@
 ### dev
 
-[full changelog](http://github.com/yolk/sum_sum/compare/v0.0.4...master)
+[full changelog](http://github.com/yolk/sum_sum/compare/v0.0.5...master)
+
+### 0.0.5 / 2011-01-27
+
+[full changelog](http://github.com/yolk/sum_sum/compare/v0.0.4...v0.0.5)
 
 * Added SumSum#level
 * Some code cleanup